+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 ## [4.0.4-beta]
 
-Small dependancy fix.
+Bug fix.
 
 ### Fixed
 - Update to fixed Poly Mixin version for Fabric 1.20.1
+- Fix floating torches (or others using support shape methods)
 
 ## [4.0.3-beta]