+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [4.0.4-beta]
+
+Repair outdated config files on game load.
+
+### Fixed
+- Update to fixed Poly Mixin version for Fabric 1.20.1 (less unstable)
+- Fix floating torches (or others using support shape methods)
+- Fix invalid config file crash
+
 ## [4.0.3-beta]
 
 Add Poly Mixin dependency for better automatic compatibility.